requirements.txt
asttokens==2.2.1
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
dash==2.8.1
dash-bootstrap-components==1.3.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
Flask==2.2.2
google-api-core==2.11.0
google-api-python-client==2.75.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
ipykernel==6.20.2
ipython==8.9.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyterlab-widgets==3.0.5
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.1
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.6.2
plotly==5.13.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
psycopg2==2.9.5
# psycopg2-binary==2.9.5
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pywin32==305; platform_system=="Windows"
pywin32-ctypes==0.2.0
pyzmq==25.0.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
stack-data==0.6.2
tenacity==8.1.0
tornado==6.2
traitlets==5.9.0
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.2
widgetsnbextension==4.0.5
# gunicorn==20.0.4